[VFX][Fix] Smoke Lighting shaders were incompatible with instancing
This PR fixes issues that Smoke Lighting had with instancing. The previous code was assuming the alphaRemap and emissiveGradient parameters where in a constant buffer, which is not the case anymore with instancing. 
These values are now passed though varyings.

The test scene and reference images are also updated to not use TAA, which was causing discrepancies between Yamato and tests ran locally.

Jira Port opened here : https://jira.unity3d.com/browse/UUM-2820