MTT-1785 Marked player settings BuildTargetGroup API as obsolete
When adding the NamedBuildTarget (https://ono.unity3d.com/unity/unity/pull-request/125925/_/platform/server/build-target) as part of the introduction of the DedicatedServer subtarget, we needed to mark the old PlayerSettings API using BuildTargetGroup as obsolete. The API updater could not be used in this case and so we had a blocker as all the usages in our packages were required to be updated first.
This PR completes the task as we coordinate with the different teams to make the update on their side and unblock us.