[HDRP] Fix compile error on certain platforms that load the denoising package
In platforms that support HW raytracing but do not support the denoising package, we were getting a compile error if the package was in the project manifest. 

With this PR we only enable the denoising code path on windows, since this is the only platform we support path tracing with denoising.

Related slack discussion: https://unity.slack.com/archives/GHD5LADU7/p1654610056852229