[URP] PostPro project: Adding a #if check for XRSettings conditional check.
Fixing a broken script in URP's Post processing project.
The script is used for a Motion Blur test but breaks GameCoreScarlett + GameCoreXBoxOne because it's trying to access XR Settings that need to be in a #if conditional as they are not available there. A jira task will be created to address this issue so we can no longer get these kind of breakages but this PR focuses on getting the CI jobs back running.