[RPW] Allow shader from HDRP and URP correctly fallback to error shader
Fix Fallback shaders for URP and HDRP when pipeline for the shader is not active.

https://jira.unity3d.com/browse/UUM-1417

The far line is HDRP, Closest 5 sphere is Builtin shaders as reference. URP line in between,
![ShaderFallback](https://media.github.cds.internal.unity3d.com/user/5932/files/2845592d-4bac-4e27-82ed-b74811f3caeb)

Beware that some materials invisible when used on spheres (Sprite Mask, VR shaders).
Some shaders fallback to the normal Builtin shaders so they remain visible.
The main check that object visible when corresponding pipeline active. It means that shaders may not be pink.