[HDRP] Add caustics support in material samples eye scene
Now that [this change](https://github.cds.internal.unity3d.com/unity/unity/pull/9238) have landed, this PR adds the support for caustics as well in the graph provided in the material sample eye scene for a better eye visual. 

![image](https://media.github.cds.internal.unity3d.com/user/1764/files/282ae766-0366-441c-b366-ace2dbdff462)


This adds 4 properties to the master stack by default: 
- two already hardcoded in the graph (Iris Plane Offset and Iris Radius)
- Caustic Multiplier (to increase / dim the effect of the caustic)
- Caustic Blend (Blend between classic eye and caustic one)

![image](https://media.github.cds.internal.unity3d.com/user/1764/files/253baae2-f368-455d-b693-b00f319a8b91)


A new section has been added on the graph to control the blend and multiplier.
Multiplier is between 0 and 8, default is 1. 
Blend is between 0 and 1, default is 1.

![image](https://media.github.cds.internal.unity3d.com/user/1764/files/9b8b5286-82ae-4dd0-abb3-f3c6a8b44231)

The scene has been slightly modified to include some text
![image](https://media.github.cds.internal.unity3d.com/user/1764/files/b0408f0c-4612-4c60-9eb6-caa2b7966870)


**Before / After**
![dcfd969655c84aaccb22e8239d5a5460](https://media.github.cds.internal.unity3d.com/user/1764/files/971e26ad-3529-4eac-ac3a-e2ce20b63e92)


Caustics with rotating dir light
![a7b6f579eaa416c98ea83f2a4cfb751c](https://media.github.cds.internal.unity3d.com/user/1764/files/8911a987-c3a2-44c9-b3ac-251c93968019)