Fix ACES approximation on switch
Previous ACES approx didn't just scale the constants of the quadratics in the numerator and denumerator, but also eliminated, by restructuring the equations, the x^2 factor. 

When the new approximation was introduced, the restructuring of the equation was not done and therefore only the scaling of the coefficient was added, leading to vastly wrong result and different equation from the non-switch version. 

With the new approx I don't think scaling coefficient is needed to avoid the float overflow and in fact might introduce new issues (e.g. the constant term in the numerator is already very small), but the removal of the x^2 might, so I did just that here. 


This should fix the newly introduced issue with Switch test on URP. 

Note for future maintainers, in case the issue with overflow appears, instead of blindly scaling all factors like was done before (by the **previous** approximation),  please only scale the ones that are affecting the acescg and not the constant ones.