[RPW]Launch the material upgrader when is really needed
So this [fogbugz](https://fogbugz.unity3d.com/f/cases/1422062/), was showing the hdrp material upgrader (and urp material upgrader, but it was crashing before). So the main issue here is that the `LoadFileAndForget` for the `HDProjectSettings `and `URPProjectSettings `was not returning/loading the file and we were creating a default project settings, where we had the material last version to -1. I am not sure why the Standalone Profiler was not properly loading, I think because the AssetDatabase is not runing in that editor mode.  So for now and until we got a better material upgrading framework. I considered the following fix until we got better AssetDatabase management API.

By checking the current pipeline we are improving two things
- The Standalone Profiler will never have a current pipeline, therefore we will never execute the material upgrading.
- The material upgrading will never be kicked until you are using the pipeline, so for instance, I tested the changes on a project with both packages installed, but URP need an upgrade. I used HDRP, nothing upgrading appeared, and then I switch to URP  the material upgrading appeared. Before directly when I opened the project was appearing to me the pop up.

Also this [JIRA](https://jira.unity3d.com/browse/UUM-2434)