[ShaderGraph][Backport][2021.3] Make BuiltIn Target honor custom shader GUI
Fix for https://fogbugz.unity3d.com/f/cases/1399808/

The BuiltIn Target was not correctly honoring a custom editor GUI the way the URP and HDRP targets do. This change properly updates the TargetSetupContext with the name of a custom editor GUI if one is entered on the Target in the graph settings. This works for both the Lit and Unlit SubTargets.