Use ordinal string comparison and avoid unnecessary calls to IndexOf
I have been profiling Gigaya importing and opening up, and I noticed that VFX code generation takes up a bunch of time doing string operations. In particular, it spents an ungodly amount of time in `String.IndexOf`. Out of a total of about 1:50m import time, it spends 20s in that particular function. There are two reasons for this: First, when used with a string argument you should use `StringComparison.Ordinal` to avoid going through the super slow codepath that tries the merge all kinds of different unicode characters to do the comparison. Second, part of the shader code generation just does three times as many `IndexOf` calls than it actually needs to do. I've fixed both, and this reduces the 20s of `String.IndexOf` to "doesn't show up in profiler anymore"

Before:
![image](https://media.github.cds.internal.unity3d.com/user/1745/files/bc847a47-8f6e-423d-8478-0b27859529f0)

After:
![image](https://media.github.cds.internal.unity3d.com/user/1745/files/e10f7fe7-62d1-4b74-8ba5-00e6b6018b01)

On a related note, you can also see that there is way too much time spent in `StringBuilder.Replace` but this is harder to fix because it would require larger rewrites. It would probably still makes sense to do it at some point, mostly because Mono's implementation of `StringBuilder.Replace` is kinda... bad? :)