Additional functionality to CoreEditorUtils and CoreEditorDrawers incl. adding menu items and extending editor foldout save states
Required functionality:
* List of ```ExpandedState```s and a list of ```AdditionalPropertiesState```s given a list of headers.
* Adding menu items in the burger menu from the ```FoldoutGroup``` and ```AdditionalPropertiesFoldoutGroup``` while still keeping additional properties settings in the burger menu.
* Changed boxed for the headers and splitters in order to fit into vertical rects.
* Added shouldUpdate to FoldoutToggleHeader in order to avoid updating when called from a serialized property which will break the editor code.

Before the isBoxed change:
![image](https://media.github.cds.internal.unity3d.com/user/2017/files/c557380c-8e9c-4570-a242-e80fa149eee5)
(Image where we can see the background rect and splitter is going outside the rect on the right side)
After the isBoxed change:
![image](https://media.github.cds.internal.unity3d.com/user/2017/files/14efa2e7-03c6-4a3f-820a-70ff3baa1381)
(Image showing new boxed to fit the rect which is from BeginVertical. The image also shows custom menu items from the burger menu. It's also the use case of not updating inside an editor when called from a property o.w. the editor would be locked only to be able to edit below the last ```FoldoutHeaderToggle```)

![image](https://media.github.cds.internal.unity3d.com/user/2017/files/3c825231-ea1f-4a7d-b499-bfc9e3f5b12b)
(Image showing added content to the menu together with the additional properties settings)

![image](https://media.github.cds.internal.unity3d.com/user/2017/files/5401bba7-4bad-4b9c-b121-0177793b8723)
(Image showing the use case of having a list of ```ExpandedState``` and a list of ```AdditionalPropertiesState``` where each renderer is a property and not an editor.)