Fixed hard eges on volumetric clouds UUM-3153
https://jira.unity3d.com/browse/UUM-3153
When fixing this bug https://fogbugz.unity3d.com/f/cases/1388279/ a regression was introduced that made the edges of the clouds hard. This PR changes the way the fix is done to avoid that side effect.
Note that we should also make sure that this issue doesn't repro again after the fix.
https://fogbugz.unity3d.com/f/cases/1407759/