[VFX][Fix 1406877] Fix collision with SDF
This PR fixes some issues with the _Collide with SDF_  block : https://fogbugz.unity3d.com/f/cases/1406977/

The issue originated from a discussion on the forum (https://forum.unity.com/threads/sdf-collision-is-leaking-particles.1240378/). The user had issue mostly because the SDF used was poorly baked (baking bounds intersecting the mesh, mesh not closed), but it also highlighted an issue where particles were colliding even when they were outside the block's Oriented Box.

On top of that, the SDF Baker is slightly modified to close the 0-isosurface of the SDF when the inside of the SDF intersects the border of the texture.