[Backport] 2022.1: Identify procedural VT properties as "per material" properties (v2)
This PR backports
- Engine PR https://ono.unity3d.com/unity/unity/pull-request/136506 (including the additional fix https://ono.unity3d.com/unity/unity/pull-request/145143)
- Package PR https://github.com/Unity-Technologies/Graphics/pull/6176

When creating multiple materials using the same shader which has a PVT-node, it's not possible to have each material bind its own procedural stacks in combination with having the SRP batcher active. PVT-nodes used to break SRP batcher compatiblity (VT rendered properly, but only because batching was not active); this compatibility was recently fixed, but it was discovered that trying to bind different stacks does not render properly, when materials are rendered *in the same batch*.

Note that, although discoverable, PVT is not publicly advertised yet, and Colossal Order is the only client actively making use of this feature. They have validated the changes and are happy with the fix.