2021.3/graphics/virtualtexturing/procedural vt per material
Backport of
- Engine PR https://ono.unity3d.com/unity/unity/pull-request/136506 (and additional fix https://ono.unity3d.com/unity/unity/pull-request/145143)
- and Package PR https://github.com/Unity-Technologies/Graphics/pull/6176.

When creating multiple materials using the same shader which has a PVT-node, it's not possible to have each material bind its own procedural stacks in combination with having the SRP batcher active. PVT-nodes used to break SRP batcher compatiblity (VT rendered properly, but only because batching was not active); this compatibility was recently fixed, but it was discovered that trying to bind different stacks does not render properly, when materials are rendered *in the same batch*.

Additionally, this PR also refactors the VT code wrt the Shader class; implementing all VT-specific parts in a separate ShaderTextureStackManager.