[RPW]Shader.globalRenderPipeline set up unification.
For cross pipeline shader stripping we need to know the current` Shader.globalRenderPipeline`  that is being used. As  this property is only set on the ctor of the `HDRenderPipeline `and on the `URPRenderPipeline`, **after the editor does a render**, we can not directly rely on this property, as a render could have not been done, therefore we need to store this information into the RPAssets that are being referenced into the `GraphicsSetting `and/ or in `QualitySettings` when doing the build.

Besides this I've also implemented the auto set of that property on the `RenderManager`

Custom SRP implementations will get a warning if they have not specified the property on

![image](https://media.github.cds.internal.unity3d.com/user/3279/files/71ab541a-c479-4887-b258-7590a849c19a)