[HDRP] More or less fix render queue UX with variants
No jira associated because it's really a minor UX fix

In HDRP we use render queue for internal stuff so we don't allow complete customization but we provide a rendering pass enum to still control some ordering.
This means we internally modify the render queue value even if it's not visible for the user. This is kinda strange with material variants, as technically it's an override but user don't really care about it.

For example, when modifying the receive decal variable it changes the render queue so it results in an override bar. Undoing the change keeps the override bar
![rq_no_fix](https://media.github.cds.internal.unity3d.com/user/2154/files/fec174dd-9e59-4709-88ba-fe4728cfcfe4)

Altough since we actually override the value we can do a special case to reset the override. This results in this which is a bit better.
![rq_with_fix](https://media.github.cds.internal.unity3d.com/user/2154/files/ffaebc64-038f-4a63-b192-e680ea8494a7)