[RPW]Avoid overlapped values in DebugUI.EnumFields
Fixing https://jira.unity3d.com/browse/UUM-6504.

With the refactors and the fixes, we no longer support overlapped values into `DebugUI.EnumFields`. The Material -> Material has overlapped values on the enum. For instance see:

| Entry Name| Index | Mapped Value |
| ----------- | ----------- | ----------- |
| AxF/Opacity (1)| ![image](https://media.github.cds.internal.unity3d.com/user/3279/files/596af02c-39c7-4e60-8588-f496657369c2) | ![image](https://media.github.cds.internal.unity3d.com/user/3279/files/784b03d4-09dd-4402-9c99-195f8caf68e0) |
| Eye/Opacity (48)| ![image](https://media.github.cds.internal.unity3d.com/user/3279/files/7a313f4e-d2db-45e7-991c-2481b7f9a75b)| ![image](https://media.github.cds.internal.unity3d.com/user/3279/files/18b7129f-4feb-45d2-b7f9-f0acc1150e47)|

This is not longer supported as the values of the enum fields must be unique.

So to workarround this, I've inserted the "Common" entries from the BuiltInData into a "Common" Section:

![image](https://media.github.cds.internal.unity3d.com/user/3279/files/403de077-0107-4a10-94e8-37c84a7a8f88)

Besides this, I've sorted the elements on the popup ( Notice that now they are alphabetical ordered).

Added a warning just in case there are users wronly using the enum field.