Hardware Shader Variants - Wave Operations
The goal of this PR is to make it easier for developers to author
shaders that target optional hardware features such as wave
operations.

This PR attempts to achieve this goal by defining a set of standard
shader keywords which are tied to specific shader requirements inside
the engine. This PR also modifies the engine to automatically set
these standard keywords based off of the current device's
capabilities.

With these changes, targeting a optional hardware feature inside a
shader is as simple as adding:

#pragma multi_compile _ UNITY_DEVICE_SUPPORTS_X

This PR also adds a few helper defines to the SRP Core Common.hlsl
file in order to simplify usage. These helpers attempt to bridge the
gap between platform features and device features for shader
authors by making them both accessible through a single define.

NOTE: This PR contains a few cleanup changes which required changes
in the build logic for CgBatch. (Needed keyword name string access) It
also modifies the logic in several of the GfxDevice backends in order to
expose a consistent set of wave operations across different devices.