[RPW] Allow shader from HDRP and URP correctly fallback to error shader, attempt 2
Fix Fallback shaders for URP and HDRP when pipeline for the shader is not active.

https://jira.unity3d.com/browse/UUM-1417
![FallbackTests](https://media.github.cds.internal.unity3d.com/user/5932/files/4887e006-239d-4e5c-851e-4b53a18c4faf)

Beware that some materials invisible when used on spheres (Sprite Mask, VR shaders).
Some shaders fallback to the normal Builtin shaders so they remain visible.
The main check that object always visible (except for vr occlusion). It means that shaders may not be pink.