[VFX][Fix 1429720] Replace default mesh output shader
This PR replaces the default shader used for the Mesh Output context, because it was outdated and hence was not compatible with the SRPs. It has been replaced with a ShaderGraph that exposes a texture and a base color, that is compatible with both URP and HDRP. (see discussion here https://unity.slack.com/archives/G1BTWN88Z/p1653037686922009)
The "template" Simple Mesh Output has also been updated to use this ShaderGraph.