[RPW]Fix FreeCamera printing an error on the first frame
This PR fixes the issue reported in https://jira.unity3d.com/browse/UUM-2697.

The issue is caused by FreeCamera doing the following:
```
void OnEnable() {
// ...
    InputRegistering.RegisterInputs(/* new input actions */);
}

void Update() {
    Input.GetAxis(/* one of the new input actions */)
}
```
This was a problem because `InputRegistering.RegisterInputs` deferred the actual registration to the next frame using `EditorApplication.delayCall`. This was fixed by removing the use of `delayCall` because it was an optimization which was determined to not be very effective anyway. Now the new inputs get registered immediately.