DGI initial state from baked (#99)
* A prototype of DGI radiance initialization from baked SH.

* Disable ZHWindow in ProbePropagationInitialize for now as it gives too smooth and dark radiance to axes.

* Minor fixes in propagation and a cleanup.

* Partner/crazyhunter 10.7.0 custom.1 dgi initial state from baked sh to from basis improvement (#107)

Improve ambient dice basis coefficients, improve basis normalization, and improve methods for projection from and to SH, choosing the right method for the context.

Dynamic GI: Ambient Dice to ZH: Switched over to using a raw ZH fit instead of deringed zh fit. It results in significantly lower error when compared to a ground truth monte carlo projection, which means slightly punchier + higher contrast + less washed out GI. And the ringing concern seems to not be an issue due to the way light is propagated. Across a number of test scenes no ringing is present.

Co-authored-by: Nicholas Brancaccio <pastasfuture@gmail.com>