[RPW]Move core Blitter initialization to URP pipeline constructor
This PR fixes the following bug: https://jira.unity3d.com/browse/UUM-2432

The issue was that URP was calling `Blitter.Initialize()` when the URP asset was selected for the first time. This is not okay because `Blitter` is a shared class between HDRP and URP, so initializing it with URP data while HDRP is active broke the blitter for HDRP, causing its Final Blit to fail.

The fix is to call `Blitter.Initialize()` only when `UniversalRenderPipeline` itself is being constructed, because then we know that it won't conflict with HDRP.