Fixed ShaderGraph import issue when interacting with render pipeline packages
The purpose of this PR is to fix ShaderGraph issues when an SRP package is installed/uninstalled.

In order to do this, we register in the ShaderGraphImporter a dependency on a hash, this hash is updated every time we change the type of the current render pipeline asset.
The hash is registered only if there is a target that couldn't be loaded during the last import in ShaderGraph, this means that some code was missing during the last import and the shader is probably not working as intended.

This work also fixes the problem when we import a render pipeline with a compilation error in the project:
- For example, when URP is imported with compilation errors, all the SG in URP are imported before the code
- This results in all URP SGs being broken as they don't have the URP Target
- With this change, they will be reimported when we switch to URP, thus fixing the broken shaders.

Note that this doesn't fix the case where you uninstall an SRP package and this operation causes the project to not compile. This issue is tracked by the asset DB team here: https://jira.unity3d.com/browse/UUM-8521