[21.3][URP] Strip default URP pipeline asset from build when all quality levels for current target have overrides
This PR is a backport of https://github.cds.internal.unity3d.com/unity/unity/pull/12566

It fixes [UUM-8511](https://jira.unity3d.com/browse/UUM-8511)

URP before Unity 23.1 always adds the default RP asset from under `ProjectSettings` -> `Graphics` to player builds. However when quality settings selected for the current build target have URP pipeline asset overrides, the default asset and it's features should be ignored.

The current behavior is not desirable and can result in overhead when the unused URP asset requests additional shaders to be compiled and packaged into the Player data (the overhead can be dramatic)


(The issue was fixed in trunk [2023.1](https://github.cds.internal.unity3d.com/unity/unity/pull/11312) after the addition of a new QualitySettings API (has been [backoprted to 22.2](https://github.cds.internal.unity3d.com/unity/unity/pull/12556) where the new API is also available))

This PR doesn't use the new API but tries to handle default assets in a similar way that HDRP did (as it did not have this issue)

After this PR the default asset will only be added to the player build when either of the following is true:
- The current build target has no quality levels selected
- The current build target has at least one quality level selected which has no valid URP asset override

This PR also makes sure that the RP assets in the quality levels and in the Graphics Settings are actually meant for URP before trying to use them (they are treated as missing otherwise)