[URP] Add "-render-graph-reuse-tests" command line option to enable RG for CI testing
We need to add RenderGraph test coverage to the CI.
All the existing projects and scenes need to run with RG enabled and reusing all the same reference images and settings.

This approach is the same used by the XR tests and is the most effective for what we need, while waiting to enable RG by default.

Jira:
https://jira.unity3d.com/browse/URP-1406