Migrate from ComputeBuffer to GraphicsBuffer Resources in Render Graph
This PR makes a change to the Render Graph API and how it manages buffer resources.

Before the PR, Render Graph buffer management was based on ComputeBuffers. 

With this PR, buffer management is now based on the more modern GraphicsBuffers, allowing for RG Render Graph interoperability with GraphicsBuffer-only APIs, like the somewhat recent Mesh API improvements (something we require for the in-dev hair rasterizer). 

I also upgraded HDRP's usage of Render Graph buffers to reflect this change.

Natively, ComputeBuffers are actually just an alias for GraphicsBuffers, so this is a pretty simple change to make. The only concern is that since this touches a public API, we may need to do an API upgrader (if that is possible for package code?).