Disabling depth priming on GLES when MSAA is enabled and when baking reflection probes
Jiras: 
https://jira.unity3d.com/browse/UUM-14181
https://jira.unity3d.com/browse/UUM-14147

Depth priming doesn't work on GLES when MSAA is enabled, because of copy depth limitations on that platform. This PR disables completely depth priming on GLES when MSAA is on to prevent users to get wrong results.
It also disables depth priming when baking reflection probes, which used to cause reflection shadow artifacts.

I also cleaned up and refactored all the depth priming logic since it was full of redundant and unneeded checks, and potential errors due to faulty logic.