[HDRP] Fix light layers when writing motion vectors
Object that would write motion vectors would not output 0 in the light layer slot, so it would disable them.
It only happens in scene view because motion vectors are not available in editor view.
It doesn't repro on trunk because rendering layers don't have this problem as light layers and decal layers are the same