2022.1: Scene is brighter in Standalone player if it was open in the Editor at build time
This is a fix for case 1375015 where environment lighting was missing from player builds in certain scenarios. The bug no longer reproduces exactly with the steps listed, but environment lighting will in general be missing when the scene hasn't been baked, and the environment lighting should therefore from the skymanager.

Backport case UUM-4703
Original case UUM-2743