[VFX] Add Maximum Capacity
This PR is following up https://github.cds.internal.unity3d.com/unity/unity/pull/11314, it solves this [issue](https://jira.unity3d.com/browse/UUM-5665) adding a maximum capacity in VFXManager. The default value is one hundred million particles and it seems to be a reasonable default limit for current revision.

The main goal of this PR is to prevent a GPU hang with really big simulation.