[URP] Shader cleanup: Removing a Subshader and some code cleanup
The goal of this PR is to:
* Reduce the build times for platforms using DirectX 11 and URP.
* Simplify URP's shader code

### This PR does the following:

#### Use include_with_pragmas for DOTS and WRITE_RENDERING_LAYERS
The PR makes use of the `#include_with_pragmas` ([Documentation](https://docs.unity3d.com/Manual/shader-include-directives.html)) shader feature and moves the relevant DOTS & Rendering Layer pragmas in to a separate file, instead of them being copy-pasted in each pass.

#### Removes a subshader in URP's default shaders
The extra subshader that was added to URP Shaders when DOTS support was implemented.

#### Removes a subshader in URP's ShaderGraph
Same as the one above except here I needed to add the the `#include_with_pragmas` ([Documentation](https://docs.unity3d.com/Manual/shader-include-directives.html)) shader feature to ShaderGraph.

#### Make all GBuffer passes have target 4.5
The target was 4.5 in URP Shaders but 2.0 or 3.0 in Terrain shaders and not specified in SpeedTree shaders.

#### Removes the 'norenderinglayer' instancing option from Terrain shaders.
The 'norenderinglayer' command is not used and generates a warning in Terrain shaders.
The code in CgBatch.cpp looks like this:
```
else if (t == "renderinglayer")
{
    out.p.compileFlags |= kShaderCompFlags_InstanceRenderingLayer;
}
```
so if the `renderinglayer` command is not there, it will not adding it to instancing. 
The `norenderinglayer` command is however never used.


#### Shader code cleanup
The shaders in URP are very inconsistent in terms of how they are setup. This PR improves the structure a bit by adding a few comments to split the code up. With this change, most of the URP shaders ares now set up like this

* Subshader
* Tags
* LOD
* Render state Commands
* Pass
    * Name
    * Tags
    * Render state Commands
    * Shader Stages
    * Material Keywords
    * Universal Pipeline keywords
    * Unity defined keywords
    * GPU Instancing
    * Includes