Shader prefiltering mvp
Adding lots of multi_compile to shaders can make the shader variant space to grow to unmaintainable proportions. Even though we do support stripping the variants using scriptable stripping, it is easily possible to have too many variants to process through the stripping system. This already affects URP build times today and therefore we have both a planned feature epic and and a high user pain bug to alleviate this problem.
https://jira.unity3d.com/browse/PLAT-1449
https://jira.unity3d.com/browse/UUM-3711

This feature introduces a renderpipeline settings based pre-filtering solution to reduce the amount of variants that we need to process early in the pipeline. It is basically a C# attribute based system where data fields in the renderpipeline asset and related scripts can be marked up to affect the shader keyword setup. See the [TDD](https://docs.google.com/document/d/1IeNrSMsYaqnTKzKBBEhEM9TP8JBdbUOTQfY863H5emA) for more details (not fully up to date with the current state though).
Looking through the added filter attributes in the URP renderpipe settings and the added API docs gives the most up-to date picture of the functionality.

Here is an example result of reduction in the variant space size when building URP foundation tests:
Compiling shader "Universal Render Pipeline/Lit" pass "ForwardLit" (fp)
Full variant space: 2013265920
After settings filtering: 18350080

That is over 100x reduction in the amount of variants that need to be processed with the following stripping stages.
This filtered variant set then goes through the scriptable stripping that reduces the compiled set even further as usual.