[2023.1] Don't unnecessarily call AssetDatabase.FindAssets
While looking at a customer project, I noticed that the VFX GRaph package shows up multiple times as an offender during domain reloads. In both cases, the problem is a call to `AssetDatabase.FindAssets`. Calling this API is _never_ a good idea. _Never_. It scales with the size of the project and potentially visits all assets. One of these places is in [VFXResources.OnEnable](https://github.cds.internal.unity3d.com/unity/unity/blob/d4545c4a8a3de83b4685daff55c718f6bd8fb9fb/Packages/com.unity.visualeffectgraph/Editor/Utils/VFXResources.cs#L47) where you query all assets for every users on all domain reloads (wasting seconds every time on large projects) to detect an error condition that likely only affects one in ten thousand users, and my suggestion is to just delete it because the cost is unreasonable. However, I'm not touching that one because that's your decision at the end of the day :)

On the other hand, you also call this in `VFXGraph.CheckCompilationVersion()` and there it is very easy to avoid the cost by moving the call into the `if` that will likely never trigger anyway. This saves 500ms on a project that I'm looking at:

![image](https://media.github.cds.internal.unity3d.com/user/1745/files/c4ca29fb-9fef-4630-b045-06ce372de649)

(My intention is to backport this as far as possible once this lands.)