UITK Search Window (AW-1020)
The goal of this PR is to convert the Search window to UITK. There were two criteria to allow this change:
1. All the features are at parity with the old window.
2. Performance is on par or better than the old window.
We think we achieved those requirements. We also made each of the components composing the window as modular as possible so they could be reused elsewhere without much effort, as long as the requirements are met. This will enable more workflows built on search to be easily done.

There is a lot of changes in this PR. Most of them come from new code written for the components of the window. However, we created some new systems to help us achieve feature parity and fix some issues we had when converting from IMGUI to UITK. Those systems are:
- The Preview Manager: we created a preview manager to help us handle previews for the Search Items when dealing with multiple preview requests for a single Search Item. Previously, the preview would be cached on a Search Item, but each component needs to draw a different preview. This would cause some issues where, for example, the Details view preview would affect the List view previews. The Preview Manager allows us to cache different previews for a single Search Item. Furthermore, it supports fetching those previews asynchronously, so we can defer fetching the previews when they are actually needed, i.e. don't fetch when scrolling super fast.
- The Search Event Manager: in IMGUI, the UI was always up to date because it was rebuilt each frame. This made it easy to stay up to date with the data. With UITK, since the UI is retained, we need a mechanism to know when something changed and update the UI. We also needed this mechanism to not add any dependencies between the different components to ensure the best modularity, and to be thread safe. Therefore we created the Search Event Manager. This system allows to send "events" of any kind with a payload, and anyone can listen and react to those events. Those events contain some information about the emitter of the event, so we can know their origin, i.e. did they come from the current window or not. This allows to react to events coming from the current window only or from any source.

We also created a base class called SearchElement, which most of our UITK components derive from. This allows us to easily have access to the current search context and window state from any component.

This PR also brings some bug fixes and improvements.

JIRA: https://jira.unity3d.com/browse/AW-1020