[VFX] Auto-reinit when modifying values that affects spawn or init context
This PR is a small quality of life feature that adds an auto reinit when values in spawner or init contexts get modified. 

Before values at spawn/init were cumbersome to tweak for some effects. In particular those with single burst because you had to reinit the effect manually to see the changes. Now it's automatic.

A prewarm can also be specified to reinit at a certain state. It defaults to 0 and it's used only if the authored VFX has not already some runtime prewarm, otherwise the VFX prewarm values are used.

You can toggle on/off the feature in the Compile Menu
The Prewarm with reinit can also be configured there
![image](https://media.github.cds.internal.unity3d.com/user/346/files/48cdf861-7b81-434f-8c90-faab27180c2f)

Prewarm for edition can be configured in VFX Preferences
![image](https://media.github.cds.internal.unity3d.com/user/346/files/31d173f6-8bbc-4447-866c-6dc23f17f84a)

Note: This is an edition only feature, it has no incindence on runtime or serialization