[HDRP] Change Gate Fit Mode default value in hdrp
Reproduced fix from this PR https://github.com/Unity-Technologies/Graphics/pull/6830
But this time made the change directly to core unity instead on relying on a workaround so that it can't get reverted by mistake.
As a result editor needs to be recompiled to test

For that, i added a graphics settings to change the default gate fit mode on new cameras that get created.
As a result that will not have any effect on already created cameras.