[2023.1] Only query MaterialGraphEditWindow instances when necessary
When making changes to a prefab in a customer project, I noticed that the `ShaderGraphAssetPostProcessor` is taking up almost 100ms, even though I'm not doing anything shader related at all. Taking a look at the asset postprocessor itself, it spends most of its time just updating the various shader graph windows that may or may not be open. Figuring out that there are no shader graph windows open is surprisingly expensive, so let's only do that once we have figured out that we actually need to do something with these windows.

![image](https://media.github.cds.internal.unity3d.com/user/1745/files/332296ae-7f35-4eb4-93bf-488f6ef65c6a)