[VFX][Fix UUM-13590] Undefined space matrices in GlobalSortKeys
Fix of UUM-13590

When in world space, the GlobalSortKey context does not send the LocalToWorld and WorldToLocal matrices to the shader. Yet, the shader was using them to set vfxLocalToWorld and vfxWorldToLocal. (see https://ono.unity3d.com/unity/unity/pull-request/143073/_/graphics/vfx/feature/instancing/instancing-matrices)

The fix consists in checking in the shader if the system is in world space, and if so, set the space matrices to the identity.