Revert #15388
In #15388 I locked the project dependencies at launch, and that causes some problems with the purchase package (see https://unity-ci.cds.internal.unity3d.com/job/17249630/logs).
This is because I tested with a wrong revision, making me believe the fix greened the pipeline but in fact it added to the failure.
So this PR is reverting the change.