2022.1/sensorsdk/add additional material support - backport
We forgot to add a define that switch towards the SensorIntegrator.hlsl for all built-in shader when we did this PR : https://github.com/Unity-Technologies/Graphics/pull/7012

We did it just for Lit and Unlit shader.  Now when our users are using other built-in shader, they get an error at runtime saying that he must define the light buffer.

Here's the JIRA ticket related to this change :  https://jira.unity3d.com/browse/MECH-1697