[VFX][Fix UUM-542] Add Point Count to Attribute from Map blocks
The `Attribute from Map` blocks consider the number of points in an attribute map to be equal to the size of the texture when computing the sampling indices. In many cases, the point count of the attribute map is smaller than the product of the texture dimensions, leading to samples where the texture values have not been set.
See https://unity.slack.com/archives/G1BTWN88Z/p1645006638732609

This PR adds a setting `Use Point Count` to this set of blocks, enabling to specify the number of point contained in the attribute map.

Note that this setting is **not available** for the Sample Modes `Sample 2DLOD` and `Sample 3DLOD`, because they do not rely on linear indexing.