[trunk] Apply LODing to DecalProjector Gizmos in SRP packages.
Use the new CalculateLOD API to reduce Gizmo rendering cost of DecalProjectors in SRPs.

The new LODing means that gizmos for close objects are as normal, but less useful features such as the directional arrow and some extra details are skipped when the object is too far away. The gizmo rendering is also skipped entirely when too far away or off screen. 

Tested in Gigaya with all of the objects in the Oasis scene selected and on screen, the time for Gizmo rendering decreased [will include data and Superluminal captures once my PC is back up and running]