[2022.2] SRP Core: Fix potential resource leak in pipeline asset
`RenderPipelineManager.CleanupRenderPipeline()` is called by
`RenderPipelineManager.HandleRenderPipelineChange()` which in turn is
called by the C++ core when you change the pipeline.

At this point the asset with the renderer is `s_CurrentPipelineAsset`.
The pipeline gets cleaned-up by the using an abstract interface. There is
no clean-up done of `s_CurrentPipelineAsset` though. It is just set to
null.

In URP, a leak happens in the renderers due to them residing in the asset.
By cleaning-up the assets, the asset can then call `DestroyRenderers()`
which frees up its resources.

This change adds a direct reference to the renderer owning asset so that
when the pipeline is cleaned, it may call `DestroyRenderers()`.

The code cannot use `currentPipeline.asset` at this point because the
`asset` member is a getter for `GraphicsSettings.currentRenderPipeline`
which is already changed at this moment in time.