[URP][Backport 2022.2] Disabling depth priming on GLES when MSAA is enabled and when baking reflection probes
Backport of:
https://ono.unity3d.com/unity/unity/pull-request/147020/_/graphics/urp/depth_priming_fixes
https://github.cds.internal.unity3d.com/unity/unity/pull/16503

Depth priming doesn't work on GLES when MSAA is enabled, because of copy depth limitations on that platform. This PR disables completely depth priming on GLES when MSAA is on to prevent users to get wrong results.
It also disables depth priming when baking reflection probes, which used to cause reflection shadow artifacts.

I also cleaned up and refactored all the depth priming logic since it was full of redundant and unneeded checks, and potential errors due to faulty logic.