Hybrid Renderer Command Buffer (#69)
* First pass at moving all hybrid renderer graphics work onto a single command buffer that HDRP submits

* More cleanup to the HybridRendererCommandBuffer API. Added toggleable immediate mode for debugging, using int IDs instead of strings for systems, and automatic profile marker generation.

* Testing a theory: Always submit the hybrid renderer command buffer, even if HDRP is not initialized or if no cameras are rendering.

Hybrid Renderer Command Buffer: Unique systems on system name and world name - it is valid for a system to push commands independantly from multiple worlds. Force an immediate mode submit of the command buffer if we encounter a case where the HDRenderPipeline has not submitted the command buffer before the system updates again - this can happen in edge cases. (#70)

Hybrid Renderer Command Buffer: Remove out-dated assert from RequestImmediateMode - the state it was guarding against is a valid state. (#71)

Make HybridRendererCommandBuffer warnings sticky so that they do not spill every frame in the event that the command buffer is not getting flushed by OnRender() in the builds.

HybridRendererCommandBuffer: Correct struct equality comparison