Disable MSAA when the water system is enabled.
This is a backport of this PR https://github.cds.internal.unity3d.com/unity/unity/pull/14894
In the current state enabling the water system and MSAA at the same time, will throw a bunch of exceptions. In the future we will want to have a fallback that works fine without nuking the performance, but for now enabling the water system will prevent the user from enabling MSAA with proper error messages (like for ray tracing).