Fix an implementation problem related to the environment lights and ray tracing.
Backport of this PR https://github.cds.internal.unity3d.com/unity/unity/pull/15265
Due to recent changes,  an additional indirection constant buffer was introduced that broke the ray tracing behavior for reflection probes. Unfortunately, due to the nature of the regression, it was not possible to capture it with our test coverage. This PR fixes the issue.