Backports of PR: 15848 and 14517
Backports for: 

https://github.cds.internal.unity3d.com/unity/unity/pull/15848
https://github.cds.internal.unity3d.com/unity/unity/pull/14517


Fixed APV ambient leaking into preview rendering and fix issue with hash of probe volume, this was a problem as the layer mask wasn't stable upon new opening of the editor. So before this PR a warning would pop up even if the probe volume was not modified.