[VFX] Update InstancingBatch test for VFX URP
Graphic test with name "InstancingBatch" in project VFX_URP is currently non-deterministic due to Sorting being disabled.
Sorting was not disabled by mistake, this shader should not be sorted. We are testing the path where we render multiple instances of a VFX with a loop, and sorting will make that impossible.
The fix was to change the blending mode to Additive instead of Alpha Blend.
For better visibility, the background was also changed to black, and reduced the amount of particles.