[VFX] Fix Get Attribute block with instancing (to 2022.2)
When spawners were reading attributes from a VFX expression on CPU, we were not using the instance index. This means that, if we have more than one instance in the same batch, they would all read the value for the first instance.
Also, for buffer remappers used in those spawners, the instance index was not used to compute the offset to the buffers.
The fixes are just providing the instance index in the right place (the default value of that parameter was 0) to access any CPU buffer.
The Jira contains some repro steps, and also affects 009_ReadAttributeInSpawner