[URP] Fixed RenderGraph Vulkan playmode tests
Fixing the issues introduced by a previous PR in the RenderGraph tests (not on QV):
https://yamato-artifactviewer.prd.cds.internal.unity3d.com/eed9b106-7559-41a6-a522-b91321867ecc%2FTest_Report_1%2FUtrTestReport1/TestReport.html

We currently have no depth resolve support on the RenderGraph path, the DepthCopyPass however was still assuming that depth resolve was enabled. This PR makes sure that we setup the copy depth pass correctly by disabling depth resolve on it when RG is enabled.

All the failing tests were rerun locally and are now green (too many for the screenshot):
![image](https://media.github.cds.internal.unity3d.com/user/2433/files/d7b05d56-c421-44a8-a824-eb4e2b8dbaca)