Fix for memory leak inside import workers when creating a preview in an HDRP project
Fix for memory leaks happening on the Asset Import Workers when a project with an HDRP Asset would create previews.

The way in which the AsyncGPUReadbackManaged code is made, relies heavily on the Editor's player loop to dispose of resources after an update of the player loop happens. This design is not compatible with the Asset Import Workers because the Editor is not ticked inside of them, and even if it were ticked, there are no windows to repaint or systems which rely on the player loop. Therefore a catch-all global callback needed to be introduced: onImportWorkerCleanupResources which is invoked upon the completion of an import inside the Asset Import Worker Process. This in turn cleans up the memory leaks which were reported in the original case.