Fixed yamato test on 2021.3 by backporting a missing fix
This PR has been made with @pavlos-mavridis and should fix the recent failures in Yamato in 2021.3 about the cropped image. 

The fix is a missing backport of this PR for pathtracer. #14690

Also, the threshold for 5005 scene has been increased ever so slightly to make the test a bit more robust.