[VFX] Fix wrong particle count before first readback
If a particle system has "mortal" particles, the number of particles used by it is only stored on the GPU.
In those cases, to get an estimation of the particle count, we use GPU readbacks every few frames. The initial value of particle count before the first readback is received **should be 0**, but currently **an uninitialized value is returned**. This is specially visible if the Debug mode for Efficiency is being used.
This PR fixes that by ensuring that the stored value is reset every time a new particle system is created. The initial value of the buffer is also set to 0, but that is not enough if we recycle instances.
Bounds can be readback as well and they could potentially have the same issue. For bounds the buffer was already properly initialized, but this PR adds some code to ensure that it is also reset when recycling instances.