[VFX] Six-way Lighting V2
This PR is a second iteration on the Six-way smoke lit material type for VFX Graph. ([JIRA](https://jira.unity3d.com/browse/VFXG-343))
The main addition is the option to adjust the lightmaps, with the goal of being able to modify the thickness appearance of the smoke.
Part of POI : https://jira.unity3d.com/browse/POI-162
List of added settings (in the inspector): 

 - Lightmap Remap Mode
 - Lightmap Remap Ranges
 - Use Emissive Channel Scale

List of added parameters (in the output, depending on the selected settings):
 - Remap From
 - Remap To 
 - Contrast Pivot
 - Contrast 
 - Remap Curve
 - Emissive Channel Scale

In addition, some fixes have been made : 
 - Artefacts when using Linear Lights 
 - Incorrect ambient lighting directions
 - Support meshes and strips
 - Soft particles with alpha premultiplied blend mode

![image](https://media.github.cds.internal.unity3d.com/user/2768/files/d01d2009-0104-451e-a8d1-e2833d51af4b)