Enable BakePipeline + LightBaker
Over the last few months we have been landing the "pseudo" modules InputExtraction (#16728), LightBaker (#14444), and BakePipeline (#17256) along with numerous orthogonal refactorings, all in separate PRs. Landing these modules separately, enabled us to keep the PRs (including this one) at a reasonable size, in order to reduce risk and encourage more thorough reviews.

In this PR we flip the switch and enable these modules in the editor. In particular, the `Lightmapping.Bake()` (and siblings) will now be routed to `BakePipeline` instead of the old `PVRRuntimeManager`. This PR turns on a lot of new code and therefore comes with some risk. In our team we have carefully evaluated the pros/cons of landing this in time for 23.1, and we decided the pros outweigh the cons, even if it involves some risk (see Risk and Testing sections below). In particular, landing this in 23.1 will have a significant positive impact on our future/upcoming work planned for 23.2.

![etr-alt-shaded-cpu-branch](https://media.github.cds.internal.unity3d.com/user/1183/files/b7831230-e2f7-4751-9c98-1748c2e04833)
_Baked with BakePipeline/LightBaker_