Fix unstable GPU Lightmapper GFX tests
We for a while seen significant uptick of unstable GPU Lightmapper graphics tests. These are all characterized by partially or fully missing lightmaps in the resulting image. 
![ezgif-5-34d1c2f5d0](https://media.github.cds.internal.unity3d.com/user/1219/files/c001d706-8da5-49d8-8c0d-34a0ff4373e4)

The failures changes across runs and sometimes with a bit of luck there are no failures. 

There were three problems identified:
1. Instability causing the first tile to be not copied back to CPU
2. Baking error on last tile depending on the scene
3. Instability causing compositing to not be executed on fully converged lightmap or not executed at all for a lightmap

The fix for 1 is a missing state variable controlling the tile transfer from GPU to CPU, this variable is not initialized in the constructor and this can cause the first tile to simply not be transferred to the CPU after baking on the GPU. Since we pack lightmaps to the lower left corner of the lightmap first and this coincides with the 0th tile, then that could explain the dark images seen in the test runs and/or missing lighting in parts of the scene. The reason this has not been flagged up by manual testing is that the offending uninitialized state variable is initialized in the Lightmapper Reset job, so when using the Editor for "live" baking the reset job will (almost) always be issued.

The fix for 2 is to add a code path for successfully exiting a bake render job when the last tile is empty. It works correctly for all other empty tiles as the code in `IncrementTilingPassNumAndUpdateBuffersIfNeeded` is skipping over these empty tiles in a loop. The issue only applies to the last tile.

The fix for 3 is trivial but was tricky to find. By adding logging to a run on CI it was evident that there is a case where compositing may be skipped for a lightmap after it is fully converged, this leads to the pipeline continuing and completing without compositing at all. So when a removal bake buffer job is launched the pipeline cannot simply skip the compositing for that lightmap, it also needs to set the `isCompositing` flag to signal to the pipeline to keep going as a compositing job is pending to be issued for a lightmap.